paths: move paths.go to new paths package
Move the paths.go file and associated tests to a new package,
also called 'paths', out of utils.
