+++ apps/glusterfs/brick_entry.go
@@ -18,6 +18,7 @@ import (
 	"github.com/heketi/heketi/executors"
 	wdb "github.com/heketi/heketi/pkg/db"
 	"github.com/heketi/heketi/pkg/glusterfs/api"
+	"github.com/heketi/heketi/pkg/paths"
 	"github.com/heketi/heketi/pkg/utils"
 	"github.com/lpabon/godbc"
 )
@@ -66,7 +67,7 @@ func NewBrickEntry(size, tpsize, poolMetadataSize uint64,
 	entry.Info.NodeId = nodeid
 	entry.Info.DeviceId = deviceid
 	entry.Info.VolumeId = volumeid
-	entry.LvmThinPool = utils.BrickIdToThinPoolName(entry.Info.Id)
+	entry.LvmThinPool = paths.BrickIdToThinPoolName(entry.Info.Id)
 	entry.UpdatePath()
 
 	godbc.Ensure(entry.Info.Id != "")
@@ -103,7 +104,7 @@ func CloneBrickEntryFromId(tx *bolt.Tx, id string) (*BrickEntry, error) {
 	entry.Info.Id = utils.GenUUID()
 	// bricks always share a thin pool with the original brick
 	if entry.LvmThinPool == "" {
-		entry.LvmThinPool = utils.BrickIdToThinPoolName(entry.Info.Id)
+		entry.LvmThinPool = paths.BrickIdToThinPoolName(entry.Info.Id)
 	}
 	// brick clones have their own lv name (not yet known)
 	entry.LvmLv = ""
@@ -212,7 +213,7 @@ func (b *BrickEntry) Create(db wdb.RODB, executor executors.Executor) error {
 
 	req := b.brickRequest(b.Info.Path, true)
 	// remove this some time post-refactoring
-	godbc.Require(req.Path == utils.BrickPath(req.VgId, req.Name))
+	godbc.Require(req.Path == paths.BrickPath(req.VgId, req.Name))
 
 	// Create brick on node
 	logger.Info("Creating brick %v", b.Info.Id)
@@ -350,7 +351,7 @@ func addSubTypeFieldFlagForBrickEntry(tx *bolt.Tx) error {
 }
 
 func (b *BrickEntry) UpdatePath() {
-	b.Info.Path = utils.BrickPath(b.Info.DeviceId, b.Info.Id)
+	b.Info.Path = paths.BrickPath(b.Info.DeviceId, b.Info.Id)
 }
 
 func (b *BrickEntry) RemoveFromDevice(tx *bolt.Tx) error {
@@ -380,7 +381,7 @@ func (b *BrickEntry) TpName() string {
 	if b.LvmThinPool != "" {
 		return b.LvmThinPool
 	}
-	return utils.BrickIdToThinPoolName(b.Info.Id)
+	return paths.BrickIdToThinPoolName(b.Info.Id)
 }
 
 // LvName returns the expected name of the lvm lv that stores
@@ -389,7 +390,7 @@ func (b *BrickEntry) LvName() string {
 	if b.LvmLv != "" {
 		return b.LvmLv
 	}
-	return utils.BrickIdToName(b.Info.Id)
+	return paths.BrickIdToName(b.Info.Id)
 }
 
 // BrickType returns the sub-type of a brick. SubType helps determine
